refactor(api): move temporary storage index to src/shared
