+++ api/index.js
@@ -5,9 +5,9 @@ import { validateEnvironmentVariables } from './src/shared/infrastructure/valida
 validateEnvironmentVariables();
 
 import { disconnect } from './db/knex-database-connection.js';
-import { temporaryStorage } from './lib/infrastructure/temporary-storage/index.js';
 import { createServer } from './server.js';
 import { learningContentCache } from './src/shared/infrastructure/caches/learning-content-cache.js';
+import { temporaryStorage } from './src/shared/infrastructure/temporary-storage/index.js';
 import { logger } from './src/shared/infrastructure/utils/logger.js';
 import { redisMonitor } from './src/shared/infrastructure/utils/redis-monitor.js';
