+++ api/scripts/certification/generate-certification-csv-results-by-session-ids.js
@@ -12,8 +12,8 @@ import lodash from 'lodash';
 const { isEmpty } = lodash;
 import { disconnect } from '../../db/knex-database-connection.js';
 import { usecases } from '../../lib/domain/usecases/index.js';
-import { temporaryStorage } from '../../lib/infrastructure/temporary-storage/index.js';
 import { learningContentCache as cache } from '../../src/shared/infrastructure/caches/learning-content-cache.js';
+import { temporaryStorage } from '../../src/shared/infrastructure/temporary-storage/index.js';
 import { logger } from '../../src/shared/infrastructure/utils/logger.js';
 
 /**
