+++ api/scripts/certification/get-user-certification-eligibility.js
@@ -7,10 +7,10 @@ dotenv.config({ path: `${__dirname}/../.env` });
 
 import { disconnect } from '../../db/knex-database-connection.js';
 import { usecases } from '../../lib/domain/usecases/index.js';
-import { temporaryStorage } from '../../lib/infrastructure/temporary-storage/index.js';
 import * as certificationBadgesService from '../../src/certification/shared/domain/services/certification-badges-service.js';
 import * as placementProfileService from '../../src/shared/domain/services/placement-profile-service.js';
 import { learningContentCache as cache } from '../../src/shared/infrastructure/caches/learning-content-cache.js';
+import { temporaryStorage } from '../../src/shared/infrastructure/temporary-storage/index.js';
 import { logger } from '../../src/shared/infrastructure/utils/logger.js';
 
 /**
