+++ api/src/certification/enrolment/domain/services/temporary-sessions-storage-for-mass-import-service.js
@@ -1,5 +1,5 @@
-import { temporaryStorage } from '../../../../../lib/infrastructure/temporary-storage/index.js';
 import { config } from '../../../../../src/shared/config.js';
+import { temporaryStorage } from '../../../../shared/infrastructure/temporary-storage/index.js';
 
 const sessionMassImportTemporaryStorage = temporaryStorage.withPrefix('sessions-mass-import:');
 import { randomUUID } from 'node:crypto';
