+++ api/src/certification/shared/domain/services/temporary-companion-storage-service.js
@@ -1,5 +1,5 @@
-import { temporaryStorage } from '../../../../../lib/infrastructure/temporary-storage/index.js';
 import { config } from '../../../../shared/config.js';
+import { temporaryStorage } from '../../../../shared/infrastructure/temporary-storage/index.js';
 
 const _getTemporaryStorage = () => temporaryStorage.withPrefix('companion:ping:');
