+++ api/src/identity-access-management/domain/services/oidc-authentication-service.js
@@ -8,10 +8,10 @@ import { Issuer } from 'openid-client';
 import { OIDC_ERRORS } from '../../../../lib/domain/constants.js';
 import { AuthenticationMethod, AuthenticationSessionContent } from '../../../../lib/domain/models/index.js';
 import { monitoringTools } from '../../../../lib/infrastructure/monitoring-tools.js';
-import { temporaryStorage } from '../../../../lib/infrastructure/temporary-storage/index.js';
 import { config } from '../../../shared/config.js';
 import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
 import { OidcError, OidcMissingFieldsError } from '../../../shared/domain/errors.js';
+import { temporaryStorage } from '../../../shared/infrastructure/temporary-storage/index.js';
 import { logger } from '../../../shared/infrastructure/utils/logger.js';
 import { DEFAULT_CLAIM_MAPPING } from '../constants/oidc-identity-providers.js';
 import { ClaimManager } from '../models/ClaimManager.js';
