+++ api/src/identity-access-management/infrastructure/repositories/user-email.repository.js
@@ -1,5 +1,5 @@
-import { temporaryStorage } from '../../../../lib/infrastructure/temporary-storage/index.js';
 import { config } from '../../../shared/config.js';
+import { temporaryStorage } from '../../../shared/infrastructure/temporary-storage/index.js';
 import { EmailModificationDemand } from '../../domain/models/EmailModificationDemand.js';
 const verifyEmailTemporaryStorage = temporaryStorage.withPrefix('verify-email:');
